+++ src/g_game.h
@@ -67,6 +67,7 @@ void G_DeferredLoadLevel(skill_t skill, int ep, int map);
 void G_LoadGame(char *name);
 
 void G_DoLoadGame(void);
+void G_DoLoadLevel(void);
 
 // Called by M_Responder.
 void G_SaveGame(int slot, char *description, char *name);